refactor(frontend): refactor simulation plots
Break up simulation plots into smaller pieces, moving the config anf plot utility functions into `features/simulation/config.js` and `features/simulation/utils.js` respectively.
